Move /delete-register-data from HomepageResource to it's own resource class - it deserves it...
